[ADDED] Distributed Message Tracing
When an incoming message contains the header `Nats-Trace-Dest` set
to a valid subject, the server will send there messages representing
events happening in each server where the message is processed.

The events will give details about ingress, subject mapping, stream
export and service imports, jetstream and egress to subscriptions
and/or routes, leafnodes and gateways. Except for a standalone server
it is expected to receive multiple trace messages for a given inbound
message.

The header `Nats-Trace-Only`, if set to `true`, will produce the
same tracing events than without that header, except that the message
will not be delivered to subscriptions or inserted in the JetStream
stream. This is usefull to see the path that a message would take
but without affecting running applications.

Note that in a mixed environment where not all servers would be
running at 2.11+, when a message is supposed to be traced only and
the remote does not understand message tracing, the message will
not be forwared and the Egress event for that remote will indicate
the reason.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>